+++ packages/@webex/internal-plugin-llm/test/unit/spec/llm.js
@@ -2,9 +2,7 @@ import MockWebex from '@webex/test-helper-mock-webex';
 import {assert} from '@webex/test-helper-chai';
 import sinon from 'sinon';
 import Mercury from '@webex/internal-plugin-mercury';
-
 import LLMService from '@webex/internal-plugin-llm';
-
 
 describe('plugin-llm', () => {
   const locusUrl = 'locusUrl';
@@ -28,31 +26,30 @@ describe('plugin-llm', () => {
         headers: {},
         body: {
           binding: 'binding',
-          webSocketUrl: 'url'
-        }
+          webSocketUrl: 'url',
+        },
       });
     });
-
 
     describe('#registerAndConnect', () => {
       it('registers connection', async () => {
         llmService.register = sinon.stub().resolves({
           body: {
             binding: 'binding',
-            webSocketUrl: 'url'
-          }
+            webSocketUrl: 'url',
+          },
         });
         assert.equal(llmService.isConnected(), false);
         await llmService.registerAndConnect(locusUrl, datachannelUrl);
         assert.equal(llmService.isConnected(), true);
       });
 
-      it('doesn\'t registers connection for invalid input', async () => {
+      it("doesn't registers connection for invalid input", async () => {
         llmService.register = sinon.stub().resolves({
           body: {
             binding: 'binding',
-            webSocketUrl: 'url'
-          }
+            webSocketUrl: 'url',
+          },
         });
         await llmService.registerAndConnect();
         assert.equal(llmService.isConnected(), false);
@@ -63,11 +60,13 @@ describe('plugin-llm', () => {
       it('registers connection', async () => {
         await llmService.register(datachannelUrl);
 
-
-        sinon.assert.calledOnceWithExactly(llmService.request, sinon.match({
-          method: 'POST',
-          url: `${datachannelUrl}`,
-        }));
+        sinon.assert.calledOnceWithExactly(
+          llmService.request,
+          sinon.match({
+            method: 'POST',
+            url: `${datachannelUrl}`,
+          })
+        );
 
         assert.equal(llmService.getBinding(), 'binding');
       });
@@ -78,8 +77,8 @@ describe('plugin-llm', () => {
         llmService.register = sinon.stub().resolves({
           body: {
             binding: 'binding',
-            webSocketUrl: 'url'
-          }
+            webSocketUrl: 'url',
+          },
         });
         await llmService.registerAndConnect(locusUrl, datachannelUrl);
         assert.equal(llmService.getLocusUrl(), locusUrl);
@@ -90,6 +89,9 @@ describe('plugin-llm', () => {
       it('disconnects mercury', async () => {
         await llmService.disconnect();
         sinon.assert.calledOnce(llmService.disconnect);
+        assert.equal(llmService.isConnected(), false);
+        assert.equal(llmService.getLocusUrl(), undefined);
+        assert.equal(llmService.getBinding(), undefined);
       });
     });
   });