+++ packages/@webex/internal-plugin-voicea/src/voicea.ts
@@ -26,30 +26,46 @@ export class VoiceaChannel extends WebexPlugin implements IVoiceaChannel {
 
   private isTranscribingEnabled: boolean;
 
-  private vmcDeviceId: string;
+  private hasSubscribedToEvents: boolean = false;
 
+  private vmcDeviceId?: string;
+
+  private eventProcessor = (e) => {
+    e;
+    this.seqNum = e.sequenceNumber + 1;
+    switch (e.data.relayType) {
+      case VOICEA_RELAY_TYPES.ANNOUNCEMENT:
+        this.vmcDeviceId = e.headers.from;
+        this.hasVoiceaJoined = true;
+        this.processAnnouncementMessage(e.data.voiceaPayload);
+        break;
+      case VOICEA_RELAY_TYPES.TRANSLATION_RESPONSE:
+        this.processCaptionLanguageResponse(e.data.voiceaPayload);
+        break;
+      case VOICEA_RELAY_TYPES.TRANSCRIPTION:
+        this.processTranscription(e.data.voiceaPayload);
+        break;
+      default:
+        break;
+    }
+  };
   /**
    * Listen to websocket messages
    */
   private listenToEvents() {
-    this.webex.internal.llm.on('event:relay.event', (e) => {
-      this.seqNum = e.sequenceNumber + 1;
-      switch (e.data.relayType) {
-        case VOICEA_RELAY_TYPES.ANNOUNCEMENT:
-          this.vmcDeviceId = e.headers.from;
-          this.hasVoiceaJoined = true;
-          this.processAnnouncementMessage(e.data.voiceaPayload);
-          break;
-        case VOICEA_RELAY_TYPES.TRANSLATION_RESPONSE:
-          this.processCaptionLanguageResponse(e.data.voiceaPayload);
-          break;
-        case VOICEA_RELAY_TYPES.TRANSCRIPTION:
-          this.processTranscription(e.data.voiceaPayload);
-          break;
-        default:
-          break;
-      }
-    });
+    if (!this.hasSubscribedToEvents) {
+      this.webex.internal.llm.on('event:relay.event', this.eventProcessor);
+      this.hasSubscribedToEvents = true;
+    }
+  }
+
+  public deregisterEvents() {
+    this.hasVoiceaJoined = false;
+    this.areCaptionsEnabled = false;
+    this.isTranscribingEnabled = false;
+    this.vmcDeviceId = undefined;
+    this.webex.internal.llm.off('event:relay.event', this.eventProcessor);
+    this.hasSubscribedToEvents = false;
   }
 
   /**
@@ -223,6 +239,7 @@ export class VoiceaChannel extends WebexPlugin implements IVoiceaChannel {
     if (this.hasVoiceaJoined || !this.webex.internal.llm.isConnected()) return;
 
     this.listenToEvents();
+
     this.webex.internal.llm.socket.send({
       id: `${this.seqNum}`,
       type: 'publishRequest',