+++ packages/@webex/internal-plugin-voicea/src/voicea.types.ts
@@ -5,9 +5,9 @@
  * Type for payload
  */
 interface AnnouncementPayload {
-  translation:{
+  translation: {
     max_languages: number;
-    allowed_languages: string[]
+    allowed_languages: string[];
   };
 
   ASR: {
@@ -16,38 +16,39 @@ interface AnnouncementPayload {
 }
 
 // eslint-disable-next-line no-undef
-type TRANSCRIPTION_TYPES = 'unknown'|
-'eva_wake' |
-'eva_thanks' |
-'eva_cancel' |
-'highlight_created'|
-'transcript_interim_results'|
-'transcript_final_result';
+type TRANSCRIPTION_TYPES =
+  | 'unknown'
+  | 'eva_wake'
+  | 'eva_thanks'
+  | 'eva_cancel'
+  | 'highlight_created'
+  | 'transcript_interim_results'
+  | 'transcript_final_result';
 
 /**
  * Class for an Transcription Object
  */
 interface Transcription {
-  start_millis:number;
-  end_millis:number;
-  text:string;
+  start_millis: number;
+  end_millis: number;
+  text: string;
   transcript_language_code: string;
-  translations:{[x:string]:string};
-  csis:number[];
+  translations: {[x: string]: string};
+  csis: number[];
   last_packet_timestamp_ms: number;
 }
 
 /**
  * Highlights
  */
- interface Highlight {
+interface Highlight {
   highlight_id: string;
   transcript: string;
   highlight_label: string;
   highlight_source: string;
-  start_millis:number;
-  end_millis:number;
-  csis:number[];
+  start_millis: number;
+  end_millis: number;
+  csis: number[];
 }
 /**
  * Type for Transcription message
@@ -55,12 +56,12 @@ interface Transcription {
 interface TranscriptionResponse {
   type: TRANSCRIPTION_TYPES;
   transcript_id: string;
-  translations?: {[x:string]:string};
+  translations?: {[x: string]: string};
   transcripts?: Transcription[];
   transcript?: Transcription;
   highlight?: Highlight;
   csis: number[];
-  data:string;
+  data: string;
   command_response: string;
 }
 /**
@@ -69,14 +70,23 @@ interface TranscriptionResponse {
 interface CaptionLanguageResponse {
   requestId: string;
   statusCode: number;
-  errorCode:number;
-  message:string;
+  errorCode: number;
+  message: string;
 }
 
 interface IVoiceaChannel {
   setSpokenLanguage: (languageCode: string) => Promise<void>;
   requestLanguage: (languageCode: string) => void;
   turnOnCaptions: () => undefined | Promise<void>;
-  toggleTranscribing: (activate:boolean) => undefined|Promise<void>;
+  toggleTranscribing: (activate: boolean) => undefined | Promise<void>;
+  deregisterEvents: () => undefined | Promise<void>;
 }
-export {AnnouncementPayload, CaptionLanguageResponse, TranscriptionResponse, Transcription, Highlight, IVoiceaChannel};
+
+export {
+  AnnouncementPayload,
+  CaptionLanguageResponse,
+  TranscriptionResponse,
+  Transcription,
+  Highlight,
+  IVoiceaChannel,
+};